pyproject.toml
[project]
name = "fluffycloud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.1",
    "pocketsphinx>=5.0.4",
    "pyaudio>=0.2.14",
    "speechrecognition>=3.14.5",
]
